set exercise dialog ui improved

+++ app/src/main/java/com/lauraalvarez/movehabits/data/model/WorkoutExercise.kt
@@ -9,7 +9,7 @@ data class WorkoutExercise(
     val exerciseName: String = "",
     val sets: Int = 0,
     val repetitions: Int = 0,
-    val weight: Int = 0,
+    val weight: Float = 0.0f,
     val durationSec: Int = 0,
     val completedSets: Int = 0,
     val completed: Boolean = false
